pyproject.toml
[project]
name = "prolog-evaluation-metrics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=7.2.0",
    "jinja2>=3.1.6",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.2",
    "llama-index-llms-ollama>=0.9.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pyswip>=0.3.3",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "sentence-transformers>=5.3.0",
    "tqdm>=4.67.3",
]
